requirements.txt
# Core dependencies
streamlit==1.22.0
langchain==0.0.184
openai==0.27.8
python-dotenv==0.21.1

# Document processing
PyPDF2==2.12.1
openpyxl==3.0.10
pandas==1.5.3

# Vector store and embeddings
faiss-cpu==1.7.4
numpy==1.23.5
# Use a pre-built wheel for tiktoken
tiktoken>=0.3.3,<0.5.0

# API and networking
aiohttp==3.10.11
requests==2.32.3

beautifulsoup4==4.9.3

# Utilities
tqdm==4.67.1
pydantic<2.0.0  # Ensure compatibility with langchain 0.0.184
python-dateutil==2.9.0.post0
pytz==2024.2
pyyaml==6.0.2

# Build dependencies
setuptools>=41.0.0
wheel>=0.33.0